move enums to misc file
