+++ adaptivetesting/models/__init__.py
@@ -4,4 +4,4 @@ from .__item_pool import ItemPool
 from .__item_selection_exception import ItemSelectionException
 from .__test_item import TestItem
 from .__test_result import TestResult
-
+from .__misc import ResultOutputFormat, StoppingCriterion
